Extract storeArticleText logic from elasticsearch extension to index.ts
